Reduce compiler warnings on RISC OS and Linux
Behaviour-preserving warning cleanup across both toolchains:

Norcroft (RISC OS):
- Stop "Undefined macro in #if" (~375 warnings): Norcroft evaluates a bare
  undefined macro even inside "defined(X) && X", so use #ifdef / nested
  #ifdef forms instead - #if __GNUC__ -> #ifdef __GNUC__ (libtelnet.h),
  nested __STDC_VERSION__ check (libtelnet.c), nested LLONG_MAX check
  (sds.c).
- Fix "actual type 'long' mismatches format '%u'": getpid() is pid_t/long,
  so print it with %lu and an (unsigned long) cast (amnuts.c, dump.c).
- Fix "cast of int to differing enum": C enum constants are int-typed, so
  the command_table[] initialiser now casts the level column to
  (enum lvl_value) explicitly (commands.h).
- Fix "use of '=' in condition context": split the assignment out of the
  if (amnuts.c).

GCC (Linux, -Wpedantic):
- Fix "ISO C forbids an empty translation unit" for riscos_crypt.c and
  riscos_posix.c (entirely #ifdef __riscos, hence empty off RISC OS) by
  adding a benign typedef in the #else branch.

Both targets still build cleanly; no functional change.

+++ src/commands/dump.c
@@ -252,7 +252,7 @@ dump_to_file(UR_OBJECT user)
         fprintf(fp, "Node name   : %s\n", amsys->uts.nodename);
         fprintf(fp, "Running on  : %s %s %s %s\n", amsys->uts.machine,
                 amsys->uts.sysname, amsys->uts.release, amsys->uts.version);
-        fprintf(fp, "Talker PID  : %u\n", getpid());
+        fprintf(fp, "Talker PID  : %lu\n", (unsigned long) getpid());
         fprintf(fp, "Booted      : %s\n", bstr);
         fprintf(fp,
                 "Uptime      : %d day%s, %d hour%s, %d minute%s, %d second%s\n",
